Adds urgent class for alerts, auto removes all alerts after 5 seconds
Added an onload function to the body that closes any open alerts after 5 seconds.

Currently there are no urgent alerts, but if one gets passed the urgent class it is Red rather than Orange.  Eventually we will want to adjust the onload to NOT target anything marked with the class urgent.  But that is a later thing.